requirements.txt
torch==2.11.0
torchvision==0.26.0
transformers==5.5.4
pillow==12.2.0

